+++ test-build.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+cd /mnt/c/Code/agentic-persona-mapping
+./build-distribution.sh